voltando a verificação se esta logado

+++ app/Http/Controllers/Api/AddressesController.php
@@ -18,7 +18,7 @@ class AddressesController extends Controller
      */
     public function index()
     {
-        $addresses = Addresses::paginate(10); // 10 produtos por página
+        $addresses = Addresses::paginate(10);
 
         if (!$addresses->isEmpty()) {
             $paginationData = $addresses->toArray();
@@ -93,18 +93,18 @@ class AddressesController extends Controller
     public function destroy(Addresses $address)
     {
 
-        // if ($address->user_id == auth()->user()->id) {
+        if ($address->user_id == auth()->user()->id) {
         $address->delete();
 
         return response()->json([
             'status' => Response::HTTP_OK,
             'mensagem' => 'Endereço deletado'
         ], Response::HTTP_OK);
-        // }
+        }
 
-        // return response()->json([
-        //     'status' => Response::HTTP_UNAUTHORIZED,
-        //     'mensagem' => 'Você não tem permissão para deletar este produto'
-        // ], Response::HTTP_UNAUTHORIZED);
+        return response()->json([
+            'status' => Response::HTTP_UNAUTHORIZED,
+            'mensagem' => 'Você não tem permissão para deletar este produto'
+        ], Response::HTTP_UNAUTHORIZED);
     }
 }
